refactor: consolidate custom method tests into mcp_test.go
